# go/blog-service/internal/model/article.go
package model

import "github.com/GordonChen13/learn-examples/go/blog-service/pkg/app"

type Article struct {
	*Model
	Title         string `json:"title"`
	Desc          string `json:"desc"`
	Content       string `json:"content"`
	CoverImageUrl string `json:"cover_image_url"`
	State         uint8  `json:"state"`
}

type ArticleSwagger struct {
	List  []*Article
	Pager app.Pager
}

func (a *Article) TableName() string {
	return "blog_article"
}